вынес некоторую валидацию входных данных в gateway
