+++ gateway/src/main/java/ru/practicum/shareit/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.booking;
+package ru.practicum.shareit.exception;
 
 import lombok.Getter;
