+++ server/src/main/java/ru/practicum/shareit/request/ItemRequestController.java
@@ -11,7 +11,6 @@ import java.util.Collection;
 @RequestMapping(path = "/requests")
 public class ItemRequestController {
     private final ItemRequestService itemRequestService;
-    private final ItemRequestUtilService itemRequestUtilService;
 
     @PostMapping
     public ItemRequestDto create(@RequestHeader("X-Sharer-User-Id") Long userId,
@@ -21,17 +20,17 @@ public class ItemRequestController {
 
     @GetMapping
     public Collection<ItemRequestDto> getAllByRequestorId(@RequestHeader("X-Sharer-User-Id") Long requestorId) {
-        return itemRequestUtilService.findAllByRequestorId(requestorId, true);
+        return itemRequestService.findAllByRequestorId(requestorId, true);
     }
 
     @GetMapping("/all")
     public Collection<ItemRequestDto> getAll(@RequestHeader("X-Sharer-User-Id") Long requestorId) {
-        return itemRequestUtilService.findAllByRequestorId(requestorId, false);
+        return itemRequestService.findAllByRequestorId(requestorId, false);
     }
 
     @GetMapping("/{requestId}")
     public ItemRequestDto getByRequestId(@RequestHeader("X-Sharer-User-Id") Long requestorId,
                                          @PathVariable Long requestId) {
-        return itemRequestUtilService.findItemRequestById(requestorId, requestId);
+        return itemRequestService.findItemRequestById(requestorId, requestId);
     }
 }
