+++ server/src/main/java/ru/practicum/shareit/request/ItemRequestServiceImpl.java
@@ -5,13 +5,17 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.shareit.exception.NotFoundException;
+import ru.practicum.shareit.item.ItemService;
 import ru.practicum.shareit.item.ItemShortDto;
 import ru.practicum.shareit.user.User;
+import ru.practicum.shareit.user.UserDto;
 import ru.practicum.shareit.user.UserMapper;
 import ru.practicum.shareit.user.UserService;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.List;
+import java.util.Map;
 
 @Service
 @Slf4j
@@ -20,6 +24,7 @@ import java.util.Collection;
 public class ItemRequestServiceImpl implements ItemRequestService {
     private final ItemRequestStorage itemRequestRepository;
     private final UserService userService;
+    private final ItemService itemService;
 
     @Transactional
     @Override
@@ -55,5 +60,25 @@ public class ItemRequestServiceImpl implements ItemRequestService {
         });
     }
 
+    @Override
+    public ItemRequestDto findItemRequestById(Long requestorId, Long itemRequestId) {
+        return findItemRequestById(requestorId, itemRequestId,
+                itemService.findAllByRequestId(itemRequestId));
+    }
 
+    @Override
+    public Collection<ItemRequestDto> findAllByRequestorId(Long requestorId, Boolean isForRequestor) {
+        User requestor = userService.checkUser(requestorId);
+        UserDto requestorDto = UserMapper.mapToUserDto(requestor);
+        Collection<ItemRequest> requests = isForRequestor
+                ? itemRequestRepository.findAllByRequestorIdOrderByCreatedDesc(requestorId)
+                : itemRequestRepository.findAllByRequestorIdNotOrderByCreatedDesc(requestorId);
+        Map<Long, List<ItemShortDto>> requestToItemsMap = itemService
+                .findAllItemsByRequestIds(requests.stream().map(ItemRequest::getId).toList());
+        return requests
+                .stream()
+                .map(request -> ItemRequestMapper.mapToItemRequestDto(request, requestorDto,
+                        requestToItemsMap.get(request.getId())))
+                .toList();
+    }
 }
